chore: upgrade nrwl to 12.3.3
