+++ .storybook-backup/tsconfig.json
@@ -1,0 +1,5 @@
+{
+  "extends": "../tsconfig.base.json",
+  "exclude": ["../**/*.spec.ts"],
+  "include": ["../**/*"]
+}
