+++ apps/api/.eslintrc.json
@@ -4,10 +4,18 @@
   "overrides": [
     {
       "files": ["*.ts", "*.tsx", "*.js", "*.jsx"],
-      "parserOptions": { "project": ["apps/api/tsconfig.*?.json"] },
+      "parserOptions": {
+        "project": ["apps/api/tsconfig.*?.json"]
+      },
       "rules": {}
     },
-    { "files": ["*.ts", "*.tsx"], "rules": {} },
-    { "files": ["*.js", "*.jsx"], "rules": {} }
+    {
+      "files": ["*.ts", "*.tsx"],
+      "rules": {}
+    },
+    {
+      "files": ["*.js", "*.jsx"],
+      "rules": {}
+    }
   ]
 }
