+++ apps/web/.storybook/main.js
@@ -1,9 +1,12 @@
 const rootMain = require('../../../.storybook/main')
 
 // Use the following syntax to add addons!
-// rootMain.addons.push('')
+// rootMain.addons.push('');
 rootMain.stories.push(
-  ...['../src/**/*.stories.mdx', '../src/**/*.stories.@(js|jsx|ts|tsx)'],
+  ...[
+    '../src/app/**/*.stories.mdx',
+    '../src/app/**/*.stories.@(js|jsx|ts|tsx)',
+  ],
 )
 
 module.exports = rootMain
