+++ apps/web/.storybook/preview.js
@@ -1,5 +1,4 @@
 import { withKnobs } from '@storybook/addon-knobs'
 import { addDecorator } from '@storybook/react'
-import '../../../.storybook/preview'
 
 addDecorator(withKnobs)
