+++ libs/frontend/shared/.storybook/main.js
@@ -1,9 +1,12 @@
-const rootMain = require('../../../.storybook/main')
+const rootMain = require('../../../../.storybook/main')
 
 // Use the following syntax to add addons!
 // rootMain.addons.push('');
 rootMain.stories.push(
-  ...['../src/**/*.stories.mdx', '../src/**/*.stories.@(js|jsx|ts|tsx)'],
+  ...[
+    '../src/lib/**/*.stories.mdx',
+    '../src/lib/**/*.stories.@(js|jsx|ts|tsx)',
+  ],
 )
 
 module.exports = rootMain
