+++ libs/frontend/shared/.storybook/preview.js
@@ -1,4 +1,3 @@
-import '../../../.storybook/preview'
 import { withKnobs } from '@storybook/addon-knobs'
 import { addDecorator } from '@storybook/react'
