+++ libs/frontend/shared/.storybook/tsconfig.json
@@ -1,7 +1,8 @@
 {
   "extends": "../tsconfig.json",
   "compilerOptions": {
-    "emitDecoratorMetadata": true
+    "emitDecoratorMetadata": true,
+    "outDir": ""
   },
   "exclude": [
     "../**/*.spec.ts",
@@ -9,5 +10,5 @@
     "../**/*.spec.tsx",
     "../**/*.spec.jsx"
   ],
-  "include": ["../src/**/*", "./*.js"]
+  "include": ["../src/**/*", "*.js"]
 }
