+++ libs/frontend/shared/tsconfig.json
@@ -14,6 +14,9 @@
     },
     {
       "path": "./.storybook/tsconfig.json"
+    },
+    {
+      "path": "./.storybook/tsconfig.json"
     }
   ]
 }
