+++ libs/frontend/shared/tsconfig.lib.json
@@ -14,6 +14,10 @@
     "**/*.stories.ts",
     "**/*.stories.js",
     "**/*.stories.jsx",
+    "**/*.stories.tsx",
+    "**/*.stories.ts",
+    "**/*.stories.js",
+    "**/*.stories.jsx",
     "**/*.stories.tsx"
   ],
   "include": [".storybook/**/*", "**/*.js", "**/*.jsx", "**/*.ts", "**/*.tsx"]
