+++ migrations.json
@@ -1,50 +1,60 @@
 {
   "migrations": [
     {
-      "version": "12.0.0-beta.0",
-      "description": "adjusts the ngcc postinstall command to just leave 'ngcc' in there. This fixes Ivy in Jest tests and Storybooks",
-      "factory": "./src/migrations/update-12-0-0/update-ngcc-postinstall",
+      "cli": "nx",
+      "version": "12.3.1",
+      "description": "Remove deprecated options and update others according to new defaults. It syncs with the v12 migration of Angular builders.",
+      "factory": "./src/migrations/update-12-3-0/update-webpack-browser-config",
       "package": "@nrwl/angular",
-      "name": "update-ngcc-postinstall"
+      "name": "update-webpack-browser-config"
     },
     {
       "cli": "nx",
-      "version": "11.6.0-beta.0",
-      "description": "Add js patterns to tsconfig.json",
-      "factory": "./src/migrations/update-11-6-0/add-js-include-11-6-0",
-      "package": "@nrwl/next",
-      "name": "add-js-include-11.6.0"
+      "version": "12.3.1",
+      "description": "Updates storybook configurations to support webpack 5",
+      "factory": "./src/migrations/update-12-3-0/update-storybook",
+      "package": "@nrwl/angular",
+      "name": "update-storybook"
     },
     {
       "cli": "nx",
-      "version": "12.0.0-beta.0",
-      "description": "Migrate tsconfig.json to allow new jsx transform to be used. Removes the need for `import React from 'react'`",
-      "factory": "./src/migrations/update-12-0-0/use-react-jsx-in-tsconfig",
-      "package": "@nrwl/react",
-      "name": "use-react-jsx-in-tsconfig-12.0.0"
+      "version": "12.3.1",
+      "description": "Migrates some rules that have changed in Angular EsLint",
+      "factory": "./src/migrations/update-12-3-0/update-angular-eslint-rules",
+      "package": "@nrwl/angular",
+      "name": "update-angular-eslint-rules"
     },
     {
+      "version": "12.1.0-beta.1",
       "cli": "nx",
-      "version": "12.0.0-beta.0",
-      "description": "Update workspace to use `@emotion/babel-plugin` instead of `@emotion/babel-preset-css-prop` to support new jsx transform",
-      "factory": "./src/migrations/update-12-0-0/update-emotion-setup",
-      "package": "@nrwl/react",
-      "name": "update-emotion-setup-12.0.0"
+      "description": "Update jest-preset-angular to version 8.4.0",
+      "factory": "./src/migrations/update-12-1-2/update-jest-preset-angular",
+      "package": "@nrwl/jest",
+      "name": "update-jest-preset-angular-8-4-0"
     },
     {
+      "version": "12.1.2-beta.1",
       "cli": "nx",
-      "version": "12.0.0-beta.0",
-      "description": "Remove @types/react-redux from package.json since react-redux installs the package automatically now",
-      "factory": "./src/migrations/update-12-0-0/remove-react-redux-types-package",
-      "package": "@nrwl/react",
-      "name": "remove-react-redux-types-package-12.0.0"
+      "description": "Replace tsConfig with tsconfig for ts-jest in jest.config.js",
+      "factory": "./src/migrations/update-12-1-2/update-ts-jest",
+      "package": "@nrwl/jest",
+      "name": "update-ts-jest-6-5-5"
     },
     {
-      "version": "11.5.3-beta.1",
-      "description": "Update react storybook lint config",
-      "factory": "./src/migrations/update-11-5-3/update-lint-ignores",
+      "version": "12.1.0-beta.7",
+      "cli": "nx",
+      "description": "Update react storybook tsconfig to avoid VSCode errors",
+      "factory": "./src/migrations/update-12-1-0/fix-storybook-tsconfig",
       "package": "@nrwl/storybook",
-      "name": "update-11-5-3"
+      "name": "update-12-1-0"
+    },
+    {
+      "version": "12.3.0-rc.1",
+      "cli": "nx",
+      "description": "Update storybook if installed and above 6 but below 6.2.7",
+      "factory": "./src/migrations/update-11-6-0/update-storybook",
+      "package": "@nrwl/storybook",
+      "name": "update-12-3-0"
     }
   ]
 }
