+++ package.json
@@ -120,7 +120,7 @@
     "js-cookie": "2.2.1",
     "jwks-rsa": "^2.0.3",
     "lodash": "4.17.21",
-    "next": "10.1.2",
+    "next": "10.2.0",
     "node-fetch": "2.6.1",
     "passport": "^0.4.1",
     "passport-jwt": "^4.0.0",
@@ -183,30 +183,30 @@
     "@mdx-js/loader": "1.6.22",
     "@nestjs/schematics": "7.2.7",
     "@nestjs/testing": "7.6.11",
-    "@nrwl/angular": "12.0.8",
-    "@nrwl/cli": "12.0.8",
-    "@nrwl/cypress": "12.0.8",
-    "@nrwl/devkit": "12.0.8",
-    "@nrwl/eslint-plugin-nx": "12.0.8",
+    "@nrwl/angular": "12.3.3",
+    "@nrwl/cli": "12.3.3",
+    "@nrwl/cypress": "12.3.3",
+    "@nrwl/devkit": "12.3.3",
+    "@nrwl/eslint-plugin-nx": "12.3.3",
     "@nrwl/gatsby": "12.0.8",
-    "@nrwl/jest": "12.0.8",
-    "@nrwl/linter": "12.0.8",
-    "@nrwl/nest": "12.0.8",
-    "@nrwl/next": "12.0.8",
-    "@nrwl/node": "12.0.8",
-    "@nrwl/nx-cloud": "12.0.2",
-    "@nrwl/nx-plugin": "12.0.8",
-    "@nrwl/react": "12.0.8",
-    "@nrwl/storybook": "12.0.8",
-    "@nrwl/tao": "12.0.8",
-    "@nrwl/web": "12.0.8",
-    "@nrwl/workspace": "12.0.8",
-    "@storybook/addon-knobs": "6.1.11",
-    "@storybook/addon-storysource": "6.1.11",
-    "@storybook/addons": "6.1.11",
-    "@storybook/api": "6.1.11",
-    "@storybook/components": "6.1.11",
-    "@storybook/react": "6.1.11",
+    "@nrwl/jest": "12.3.3",
+    "@nrwl/linter": "12.3.3",
+    "@nrwl/nest": "12.3.3",
+    "@nrwl/next": "12.3.3",
+    "@nrwl/node": "12.3.3",
+    "@nrwl/nx-cloud": "12.1.1",
+    "@nrwl/nx-plugin": "12.3.3",
+    "@nrwl/react": "12.3.3",
+    "@nrwl/storybook": "12.3.3",
+    "@nrwl/tao": "12.3.3",
+    "@nrwl/web": "12.3.3",
+    "@nrwl/workspace": "12.3.3",
+    "@storybook/addon-knobs": "6.2.9",
+    "@storybook/addon-storysource": "6.2.9",
+    "@storybook/addons": "6.2.9",
+    "@storybook/api": "6.2.9",
+    "@storybook/components": "6.2.9",
+    "@storybook/react": "6.2.9",
     "@svgr/webpack": "5.5.0",
     "@tailwindcss/postcss7-compat": "2.1.0",
     "@testing-library/cypress": "7.0.6",
@@ -262,6 +262,7 @@
     "apollo-graphql": "0.9.2",
     "autoprefixer": "9",
     "babel-jest": "26.6.3",
+    "babel-loader": "8.1.0",
     "babel-plugin-import": "1.13.3",
     "babel-plugin-macros": "3.0.1",
     "babel-plugin-module-resolver": "4.0.0",
@@ -318,7 +319,7 @@
     "prettier": "2.2.1",
     "sass-loader": "11.0.1",
     "tailwindcss": "npm:@tailwindcss/postcss7-compat",
-    "ts-jest": "26.4.4",
+    "ts-jest": "26.5.5",
     "ts-node": "9.1.1",
     "tsconfig-paths-webpack-plugin": "3.5.1",
     "typedoc": "0.20.0-beta.26",
