+++ scripts/postinstall.sh
@@ -1,5 +1,5 @@
 #!/usr/bin/env bash
 
-if [ "$CI" != true ]; then
-  yarn codegen
-fi
+# if [ "$CI" != true ]; then
+#   yarn codegen
+# fi
