+++ workspace.json
@@ -187,6 +187,7 @@
         },
         "build-storybook": {
           "builder": "@nrwl/storybook:build",
+          "outputs": ["{options.outputPath}"],
           "options": {
             "uiFramework": "@storybook/react",
             "outputPath": "dist/storybook/web",
@@ -198,8 +199,7 @@
             "ci": {
               "quiet": true
             }
-          },
-          "outputs": ["{options.outputPath}"]
+          }
         }
       }
     },
@@ -422,9 +422,10 @@
         },
         "build-storybook": {
           "builder": "@nrwl/storybook:build",
+          "outputs": ["{options.outputPath}"],
           "options": {
             "uiFramework": "@storybook/react",
-            "outputPath": "dist/storybook/frontend",
+            "outputPath": "dist/storybook/frontend-shared",
             "config": {
               "configFolder": "libs/frontend/shared/.storybook"
             }
